Infinite scrolling implementation for booklist
The algorithm initially loads a batch of 10 books on the page. As the user scrolls, once they reach the bottom, the next batch of 10 books will render. This continues until the user reaches the end of the list (or if list is endless, user can scroll endlessly, like on social media platforms).